Work around Gtk warning
Something about Python not holding a reference to an image (which
doesn't seem to happen in any other case and I've no idea why we should
be holding a reference when it's something only Gtk itself needs ... we
don't need to do this in the case of any other images). Oh well.