Fix #3
This introduces a classic PID-file lock (although not in the classic
location as that's root accessible only) and has the SenseHat class
auto-spawn sense_emu_gui by default if nothing is detected holding the
lock. Still need to test this on Windows (good lord, detecting PIDs on
Windows is a pain).

Also bumps sense-emu-tools to "Recommends" in debian/control as there's
now a pretty strong dependency on that from the library itself, but
adding a full dep would be circular (possible but discouraged).