Fix #19
Not quite as trivial as I thought because there's two possible cases
when the simulation is disabled: either we're on a slow platform and we
still need to derive gyro, accel, etc. (with a timeout to ensure we
don't saturate the CPU when a slider is dragged), or it's during a
replay in which case we mustn't derive gyro, accel, etc. Anyway - I
think this should fix things. Time to test on some small platforms...